# Restore Julia's number type optimization from #171