Reverts the removal by @MasonProtter in #207

Also removes trailing whitespace