@cyrusmsk 2023-10-28
Btw what is the logic behind this optimization?
Smaller type will require less memory, but why it should work faster?