@MasonProtter 2023-10-28
If you do this, you'll need to change the precompilation workload to make sure each path is hit. 

I’m skeptical that there's any performance advantage though.