@MasonProtter 2023-10-28
Main:
```
[| Julia | 406.00 ms  |]
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 19 milliseconds
        Processing time (w/o IO): 20 milliseconds
        Processing time (w/o IO): 19 milliseconds
        Processing time (w/o IO): 20 milliseconds
        Processing time (w/o IO): 20 milliseconds
        Processing time (w/o IO): 20 milliseconds
        Processing time (w/o IO): 20 milliseconds
        Processing time (w/o IO): 19 milliseconds
        Processing time (w/o IO): 20 milliseconds
        Processing time (w/o IO): 20 milliseconds
        Processing time (w/o IO): 20 milliseconds
        Processing time (w/o IO): 20 milliseconds
        Processing time (w/o IO): 20 milliseconds
          Time (mean ± σ):     558.8 ms ±  10.6 ms    [User: 550.3 ms, System: 365.2 ms]
          Range (min … max):   544.3 ms … 579.9 ms    10 runs
         
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 249 milliseconds
        Processing time (w/o IO): 249 milliseconds
        Processing time (w/o IO): 257 milliseconds
          Time (mean ± σ):     855.8 ms ±  33.6 ms    [User: 864.5 ms, System: 347.9 ms]
          Range (min … max):   832.1 ms … 879.6 ms    2 runs
         
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 2188 milliseconds
        Processing time (w/o IO): 2217 milliseconds
        Processing time (w/o IO): 2297 milliseconds
          Time (mean ± σ):      3.037 s ±  0.050 s    [User: 3.008 s, System: 0.371 s]
          Range (min … max):    3.001 s …  3.072 s    2 runs
```
This PR:
```
[| Julia | 552.89 ms  |]
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 29 milliseconds
        Processing time (w/o IO): 29 milliseconds
        Processing time (w/o IO): 28 milliseconds
        Processing time (w/o IO): 33 milliseconds
        Processing time (w/o IO): 29 milliseconds
        Processing time (w/o IO): 28 milliseconds
        Processing time (w/o IO): 28 milliseconds
        Processing time (w/o IO): 28 milliseconds
        Processing time (w/o IO): 29 milliseconds
        Processing time (w/o IO): 28 milliseconds
        Processing time (w/o IO): 28 milliseconds
        Processing time (w/o IO): 28 milliseconds
        Processing time (w/o IO): 28 milliseconds
          Time (mean ± σ):     572.0 ms ±  25.2 ms    [User: 574.0 ms, System: 351.6 ms]
          Range (min … max):   551.5 ms … 634.0 ms    10 runs
         
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 334 milliseconds
        Processing time (w/o IO): 354 milliseconds
        Processing time (w/o IO): 349 milliseconds
          Time (mean ± σ):     950.1 ms ±  18.1 ms    [User: 923.2 ms, System: 383.0 ms]
          Range (min … max):   937.3 ms … 962.8 ms    2 runs
         
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 2999 milliseconds
        Processing time (w/o IO): 2986 milliseconds
        Processing time (w/o IO): 3110 milliseconds
          Time (mean ± σ):      3.848 s ±  0.092 s    [User: 3.802 s, System: 0.397 s]
          Range (min … max):    3.783 s …  3.914 s    2 runs
```