@jinyus 2023-10-28
Main:
`Julia | 37.15 ms | 481.67 ms | 4.22 s`

This PR:
```
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 34 milliseconds
        Processing time (w/o IO): 33 milliseconds
        Processing time (w/o IO): 34 milliseconds
        Processing time (w/o IO): 34 milliseconds
        Processing time (w/o IO): 34 milliseconds
        Processing time (w/o IO): 34 milliseconds
        Processing time (w/o IO): 36 milliseconds
        Processing time (w/o IO): 34 milliseconds
        Processing time (w/o IO): 35 milliseconds
        Processing time (w/o IO): 34 milliseconds
        Processing time (w/o IO): 34 milliseconds
        Processing time (w/o IO): 34 milliseconds
        Processing time (w/o IO): 34 milliseconds
          Time (mean ± σ):      1.038 s ±  0.018 s    [User: 0.911 s, System: 0.231 s]
          Range (min … max):    1.018 s …  1.065 s    10 runs
         
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 346 milliseconds
        Processing time (w/o IO): 345 milliseconds
        Processing time (w/o IO): 343 milliseconds
          Time (mean ± σ):      1.480 s ±  0.005 s    [User: 1.298 s, System: 0.285 s]
          Range (min … max):    1.476 s …  1.484 s    2 runs
         
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 2918 milliseconds
        Processing time (w/o IO): 2922 milliseconds
        Processing time (w/o IO): 2916 milliseconds
          Time (mean ± σ):      4.451 s ±  0.010 s    [User: 4.200 s, System: 0.347 s]
          Range (min … max):    4.444 s …  4.458 s    2 runs
```